Struct matrix_sdk_sqlite::SqliteStateStore

pub struct SqliteStateStore { /* private fields */ }

A sqlite based cryptostore.

Implementations

impl SqliteStateStore

pub async fn open( path: impl AsRef<Path>, passphrase: Option<&str> ) -> Result<Self, OpenStoreError>

Open the sqlite-based state store at the given path using the given passphrase to encrypt private data.

pub async fn open_with_pool( pool: SqlitePool, passphrase: Option<&str> ) -> Result<Self, OpenStoreError>

Create a sqlite-based state store using the given sqlite database pool. The given passphrase will be used to encrypt private data.