Struct matrix_sdk_sqlite::SqliteCryptoStore

pub struct SqliteCryptoStore { /* private fields */ }

A sqlite based cryptostore.

Implementations

impl SqliteCryptoStore

pub async fn open( path: impl AsRef<Path>, passphrase: Option<&str> ) -> Result<Self, OpenStoreError>

Open the sqlite-based crypto store at the given path using the given passphrase to encrypt private data.

pub async fn open_with_pool( pool: SqlitePool, passphrase: Option<&str> ) -> Result<Self, OpenStoreError>

Create a sqlite-based crypto store using the given sqlite database pool. The given passphrase will be used to encrypt private data.